==> 3 feature_engineering.ipynb <==
# Step 3: Feature Engineering

## 1. Use cleaning rules to process numeric and non-numeric data

raw_non_numeric_data.csv --> `non_numeric.py` --> processed_non_numeric_data.csv

raw_numeric_data.csv --> `numeric.py` --> processed_numeric_data.csv

## 2. Merge dataframes

import polars as pl
import os

numeric_data_path = os.path.join(os.getcwd(), "outputs", "processed_numeric_data.csv")
non_numeric_data_path = os.path.join(os.getcwd(), "outputs", "processed_non_numeric_data.csv")

# Read data
df_numeric = pl.read_csv(numeric_data_path, infer_schema=True)
df_non_numeric = pl.read_csv(non_numeric_data_path, infer_schema=True)
print(df_numeric.shape, df_non_numeric.shape)

# Join datasets on 匹配ID_日期
df = df_numeric.join(df_non_numeric, on=df_numeric.columns[0], how="inner")
df.shape

## 3. Save Dataframe

output_path = os.path.join(os.getcwd(), "outputs", "Processed_Data.csv")

df.write_csv(output_path)
